Moved Manager classes into their own namespace
